FEATURE: Add TrainingData Generator, that uses the same chipper as the inference pipeline FEATURE: Update ML Model to fastai2 FEATURE: Make CropBoxGenerator avoid producing overhangs
Docstrings
XXX WARNING Make sure to set config.UPDATE_ML = True for one run so that your local copy of the model pkl updates to the new model
Update to the new fastai2 model
Update training generator fo loop through prods
Start creating Trainingdata
Comments +training_generator

+++ ceruleanserver/ml.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 from datetime import datetime
-from fastai.vision import load_learner, get_image_files, open_image, Image
+from fastai2.learner import load_learner
 from osgeo import gdal
 from numpy import zeros, floor, ceil, uint8, where
 from PIL import Image as PILimage
@@ -23,7 +23,7 @@ def machine(learner, snso):
     reduced_chip_size = 512 # px square reduced resolution of chip_size 
     max_chip_qty = 9999
     threshold = None # Between 0 (most) and 1 (least inclusive), which confidence should be considered oil
-    start_over = True
+    start_over = True # Do we delete all files before creating chips, or pick up where we left off
     img_path = Path(snso.s3["grd_tiff_dest"])
     mask_dir = img_path.parent/'masks'
     merged_path = img_path.parent/'merged.tif'
@@ -36,7 +36,7 @@ def machine(learner, snso):
         # Should we delete all the other intermediate files at this point?
     # Store Polygons in DB
 
-def crop_box_gen(px_wide, px_high, chip_size, geo_transform=(0,1,0,0,0,1)):
+def crop_box_gen(px_wide, px_high, chip_size, geo_transform=(0,1,0,0,0,1), overhang=False):
     """A generator that will sequentially return bounding corners of boxes to chop up a larger image into small ones
     
     Arguments:
@@ -57,21 +57,21 @@ def crop_box_gen(px_wide, px_high, chip_size, geo_transform=(0,1,0,0,0,1)):
     for i in range(num_wide):
         for j in range(num_high):
             # These special cases are to avoid fractional images being processed, as they are not as likely to correctly interpret bilge slicks
-            if i==num_wide-1 and j==num_high-1:
+            if not overhang and i==num_wide-1 and j==num_high-1:
                 # very special case (corner)
                 yield [
                     geo_transform[0]+(px_wide-chip_size)*geo_transform[1]+(px_high-chip_size)*geo_transform[2], # ULX: X_origin + box_number * num_pixels_per_box * size_of_X_pixels + 0
                     geo_transform[3]+(px_wide-chip_size)*geo_transform[4]+(px_high-chip_size)*geo_transform[5], # ULY: Y_origin + 0 + box_number * num_pixels_per_box * size_of_Y_pixels
                     geo_transform[0]+px_wide*geo_transform[1]+px_high*geo_transform[2], # LRX
                     geo_transform[3]+px_wide*geo_transform[4]+px_high*geo_transform[5]] # LRY
-            elif i==num_wide-1:
+            elif not overhang and i==num_wide-1:
                 # special case (edge)
                 yield [
                     geo_transform[0]+(px_wide-chip_size)*geo_transform[1]+j*chip_size*geo_transform[2], # ULX: X_origin + box_number * num_pixels_per_box * size_of_X_pixels + 0
                     geo_transform[3]+(px_wide-chip_size)*geo_transform[4]+j*chip_size*geo_transform[5], # ULY: Y_origin + 0 + box_number * num_pixels_per_box * size_of_Y_pixels
                     geo_transform[0]+px_wide*geo_transform[1]+(j+1)*chip_size*geo_transform[2], # LRX
                     geo_transform[3]+px_wide*geo_transform[4]+(j+1)*chip_size*geo_transform[5]] # LRY
-            elif j==num_high-1:
+            elif not overhang and j==num_high-1:
                 # special case (edge)
                 yield [
                     geo_transform[0]+i*chip_size*geo_transform[1]+(px_high-chip_size)*geo_transform[2], # ULX: X_origin + box_number * num_pixels_per_box * size_of_X_pixels + 0
@@ -96,26 +96,44 @@ def create_chips(img_path, chip_size, reduced_chip_size, mask_dir, mllearner, th
         chip_dir {Path} -- Folder where the chopped up PNGs will be stored
         max_chip_qty {int} -- Optional input to reduce computational load for testing
     """
-    if mask_dir.exists() and start_over:
+    if mask_dir.exists() and start_over: # Delete any masks from a previous attempt
         shutil.rmtree(mask_dir)
-    mask_dir.mkdir(exist_ok=True)
-    tiff = gdal.Open(str(img_path))
-    boxes = crop_box_gen(tiff.RasterXSize, tiff.RasterYSize, chip_size, tiff.GetGeoTransform())
-    s = datetime.now()
+    mask_dir.mkdir(exist_ok=True) # Create new directory for masks to be stored
+    tiff = gdal.Open(str(img_path)) # Open the Large GeoTiff as a GDAL DataSet
+    boxes = crop_box_gen(tiff.RasterXSize, tiff.RasterYSize, chip_size, tiff.GetGeoTransform()) # A list of pairs of corners (UL & LR) which will be cut from the large GeoTiff
+    s = datetime.now() # Measure Inference time
     for i, b in enumerate(boxes): # 6.7s per 512px chip
-        if i<max_chip_qty:
-            png_path = mask_dir/f'{img_path.stem}_{i}.png'
-            if not png_path.exists() or start_over:
-                gdal.Translate(str(png_path), tiff, format="PNG", width=reduced_chip_size, height=reduced_chip_size, projWin=b) # check out more args here: https://gdal.org/python/osgeo.gdal-module.html#TranslateOptions
-                infer(png_path, mllearner, threshold)
+        if i<max_chip_qty: # Stop short if artificially avoiding too many chips
+            png_path = mask_dir/f'{img_path.stem}_{i}.png' # Create indexed name for new png to be stored
+            if not png_path.exists() or start_over: # Avoid overwriting images already made
+                gdal.Translate(str(png_path), tiff, format="PNG", width=reduced_chip_size, height=reduced_chip_size, projWin=b, outputType=gdal.GDT_Byte) # check out more args here: https://gdal.org/python/osgeo.gdal-module.html#TranslateOptions
+                infer(png_path, mllearner, threshold) # Run Inference on the current chip
                 if config.VERBOSE: print(png_path.stem, (datetime.now()-s)/(i+1))
     del tiff
 
 def infer(png_path, learner, threshold = None):
-    _, _, pred_class  = learner.predict(open_image(png_path)) # Currently returns classes [not bilge, bilge]
+    """Run inference on a chip
+
+    Arguments:
+        png_path {Path} -- Location of a single chip
+        learner {fastai2 model} -- Loaded fastai2 pkl file
+
+    Keyword Arguments:
+        threshold {float} -- Round predicted confidence above thresh to 1 and below to 0 (default: {None})
+    """    
+    _, _, pred_class  = learner.predict(png_path) # Currently returns classes [not bilge, bilge, vessel]
     mask = where(pred_class[1]>threshold, 1, 0) if threshold else pred_class[1]
     PILimage.fromarray(uint8(mask*255)).save(png_path)
     gdal.Warp(destNameOrDestDS=str(png_path.with_suffix('.tif')), srcDSOrSrcDSTab=str(png_path), srcNodata=0, dstNodata=0, format='GTiff')
+
+    # gdal.Translate(str(png_path.with_suffix('.tif')), str(png_path), format="GTiff")
+
+    # tif = gdal.Open(str(png_path), gdal.GA_Update)
+    # outBand = tif.GetRasterBand(1)
+    # gdal_array.BandWriteArray(outBand, pred_class[1].numpy())
+    # outBand.WriteArray(pred_class[1].numpy())
+    # outBand.FlushCache()
+    # del tif
 
 def merge_masks(mask_dir, merged_path):
     """Merge multiple chips into one large tif
@@ -126,20 +144,27 @@ def merge_masks(mask_dir, merged_path):
     """
     if config.VERBOSE: print('Merging Masks')
     if (merged_path).exists():
-        (merged_path).unlink()
-    cmd = f'gdal_merge.py -o {merged_path} {mask_dir}/*.tif -n 0'#{" ".join([str(t) for t in tif_list])}'
-    os.system(cmd) # XXX WARNING Does this limit the length of cmd by # chars?! 
+        (merged_path).unlink() # Delete previously file from previous merge attempt
+    cmd = f'gdal_merge.py -o {merged_path} {mask_dir}/*.tif'
+    os.system(cmd)
     shutil.rmtree(mask_dir)
 
 def load_learner_from_s3():
+    """Import the latest trained model from S3
+
+    Returns:
+        Learner -- fastai2 trained model loaded into memory
+    """    
     if config.VERBOSE: print('Loading Learner')
     pkl_name = 'ml.pkl'
     pkl_dir = Path('./temp/')
-    if config.UPDATE_ML:
+    if config.UPDATE_ML and (pkl_dir/pkl_name).exists():
         (pkl_dir/pkl_name).unlink() # pylint: disable=no-member
     if not (pkl_dir/pkl_name).exists(): # pylint: disable=no-member
         download_str = f'aws s3 cp {config.ML_PKL} {pkl_dir/pkl_name}'
         # print(download_str)
         os.system(download_str)
-    l = load_learner(pkl_dir, file=pkl_name)
+    l = load_learner(pkl_dir/pkl_name)
     return l
+
+def get_lbls(): return # Required by fastai as it was used during creation of the model
